refactor(app): extract mask image sizing hook
